# openapi.json
{
  "components": {
    "schemas": {
      "AppendedNoteResponse": {
        "properties": {
          "appended_bytes": {
            "description": "Bytes added to the file, including any inserted separator/terminator.",
            "title": "Appended Bytes",
            "type": "integer"
          },
          "id": {
            "description": "Vault-relative path of the note.",
            "title": "Id",
            "type": "string"
          },
          "modified_at": {
            "description": "Modification time in the configured timezone.",
            "format": "date-time",
            "title": "Modified At",
            "type": "string"
          },
          "path": {
            "description": "Vault-relative path of the note.",
            "title": "Path",
            "type": "string"
          }
        },
        "required": [
          "id",
          "path",
          "modified_at",
          "appended_bytes"
        ],
        "title": "AppendedNoteResponse",
        "type": "object"
      },
...
      "CreatedNoteResponse": {
        "properties": {
          "id": {
            "description": "Vault-relative path of the created note.",
            "title": "Id",
            "type": "string"
          },
          "modified_at": {
            "description": "Creation time in the configured timezone.",
            "format": "date-time",
            "title": "Modified At",
            "type": "string"
          },
          "path": {
            "description": "Vault-relative path of the created note.",
            "title": "Path",
            "type": "string"
          },
          "related_notes_linked": {
            "description": "Number of export.related_notes candidates that were verified and rendered as wikilinks. Always 0 for the raw content/frontmatter path.",
            "minimum": 0.0,
            "title": "Related Notes Linked",
            "type": "integer"
          },
          "related_notes_skipped": {
            "description": "Number of export.related_notes candidates omitted because they could not be verified — invalid, missing, or duplicate. Always 0 for the raw content/frontmatter path. Submitting more than the documented maximum is a separate, harder failure: the whole request is rejected before this count is ever produced.",
            "minimum": 0.0,
            "title": "Related Notes Skipped",
            "type": "integer"
          },
          "title": {
            "description": "Sanitised file-name stem of the created note, without .md. May differ from the H1 and frontmatter title of a structured export.",
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "path",
          "title",
          "modified_at",
          "related_notes_linked",
          "related_notes_skipped"
        ],
        "title": "CreatedNoteResponse",
        "type": "object"
      },
...
      "DuplicateCandidatesResponse": {
        "properties": {
          "candidate_count": {
            "description": "Total number of candidates found before `limit` was applied. `recommendation` is decided from this full set, not from `candidates`.",
            "minimum": 0.0,
            "title": "Candidate Count",
            "type": "integer"
          },
          "candidates": {
            "description": "Up to `limit` candidates, most confident first.",
            "items": {
              "$ref": "#/components/schemas/DuplicateCandidate"
            },
            "title": "Candidates",
            "type": "array"
          },
          "recommendation": {
            "description": "'create': no high/medium candidate — proceed without asking. 'confirm': exactly one high candidate and nothing else at high/medium — ask the user to choose new/append/cancel. 'choose': more than one high/medium candidate, or only medium ones — show the candidate list and require an explicit pick. This is advisory, not write authorization: the Gateway never blocks or gates create_inbox_note/append_inbox_note on this value.",
            "enum": [
              "create",
              "confirm",
              "choose"
            ],
            "title": "Recommendation",
            "type": "string"
          },
          "scanned_count": {
            "description": "Notes directly inside the inbox that were read.",
            "minimum": 0.0,
            "title": "Scanned Count",
            "type": "integer"
          },
          "skipped_count": {
            "description": "Notes directly inside the inbox that were excluded: unreadable frontmatter, or a file name that would not be accepted as an append_inbox_note target.",
            "minimum": 0.0,
            "title": "Skipped Count",
            "type": "integer"
          },
          "truncated": {
            "description": "True when `candidate_count` exceeds the number of `candidates` returned.",
            "title": "Truncated",
            "type": "boolean"
          }
        },
        "required": [
          "candidates",
          "candidate_count",
          "truncated",
          "recommendation",
          "scanned_count",
          "skipped_count"
        ],
        "title": "DuplicateCandidatesResponse",
        "type": "object"
      },
      "ErrorCode": {
        "description": "The error code vocabulary (IMPLEMENTATION_PLAN section 13).\n\nTwo codes need a note:\n\n* ``RATE_LIMITED`` is part of the published contract but is never raised in\n  Phase 1 — no rate limiting is implemented yet.\n* ``VALIDATION_ERROR`` is an addition to section 13's list (which the plan\n  calls \"主なエラーコード\", i.e. non-exhaustive). Request validation failures\n  that are not about a path or a title — a non-integer ``limit``, an unknown\n  body field — need a code of their own rather than being forced into\n  ``INVALID_PATH``.",
        "enum": [
          "UNAUTHORIZED",
          "INVALID_PATH",
          "PATH_OUTSIDE_VAULT",
          "NOTE_NOT_FOUND",
          "INVALID_FILE_TYPE",
          "FILE_TOO_LARGE",
          "INVALID_TITLE",
          "NOTE_ALREADY_EXISTS",
          "VALIDATION_ERROR",
          "RATE_LIMITED",
          "INVALID_CURSOR",
          "NOTE_MODIFIED",
          "INBOX_LOCK_TIMEOUT",
          "INTERNAL_ERROR"
        ],
        "title": "ErrorCode",
        "type": "string"
      },
      "ErrorDetail": {
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorResponse": {
        "description": "The single error shape used by every failing response (section 13).",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "InboxNoteAppendRequest": {
        "additionalProperties": false,
        "properties": {
          "content": {
            "description": "Markdown to append. The existing note's line ending is preserved.",
            "maxLength": 2000000,
            "minLength": 1,
            "title": "Content",
            "type": "string"
          },
          "path": {
            "description": "Vault-relative path of an existing .md note directly inside the inbox.",
            "maxLength": 1024,
            "minLength": 1,
            "title": "Path",
            "type": "string"
          }
        },
        "required": [
          "path",
          "content"
        ],
        "title": "InboxNoteAppendRequest",
        "type": "object"
      },
      "InboxNoteCreateRequest": {
        "additionalProperties": false,
        "description": "Exactly one of `content` or `export` is required; sending both, or\nsending neither, is rejected. `export` and `frontmatter` are also\nmutually exclusive — `export`'s frontmatter (title/created/updated/\nsource/export_mode/tags) is formatter-owned and cannot be supplied as\narbitrary free-form frontmatter alongside it. `content`/`frontmatter`\nremain the raw-Markdown path kept for existing callers; `export` is the\nstructured chat-export path (issue #12 / docs/adr/0005-*.md).",
        "properties": {
          "content": {
            "anyOf": [
              {
                "maxLength": 2000000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Markdown body. Written as-is with LF line endings. Exactly one of `content` or `export` is required.",
            "title": "Content"
          },
          "export": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ChatExport"
              },
              {
                "type": "null"
              }
            ],
            "description": "Structured chat-export input (issue #12). The Gateway renders this into deterministic Markdown and frontmatter; the client never supplies raw Markdown or frontmatter alongside it. Exactly one of `content` or `export` is required."
          },
          "frontmatter": {
            "anyOf": [
              {
                "additionalProperties": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "integer"
                    },
                    {
                      "type": "number"
                    },
                    {
                      "type": "boolean"
                    },
                    {
                      "items": {
                        "anyOf": [
                          {
                            "type": "string"
                          },
                          {
                            "type": "integer"
                          },
                          {
                            "type": "number"
                          },
                          {
                            "type": "boolean"
                          },
                          {
                            "type": "null"
                          }
                        ]
                      },
                      "type": "array"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional YAML frontmatter for the `content` path. Scalars and flat lists of scalars only. Rejected together with `export`, whose frontmatter is formatter-owned.",
            "title": "Frontmatter"
          },
          "title": {
            "description": "Human-readable title. The file name is derived from it by the API; callers cannot choose a path.",
            "maxLength": 300,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title"
        ],
        "title": "InboxNoteCreateRequest",
        "type": "object"
      },
      "NoteResponse": {
        "properties": {
          "content": {
            "description": "Markdown body with the frontmatter block removed.",
            "title": "Content",
            "type": "string"
          },
          "frontmatter": {
            "additionalProperties": true,
            "description": "Parsed YAML frontmatter. Empty when absent or unparseable.",
            "title": "Frontmatter",
            "type": "object"
          },
          "id": {
            "description": "Vault-relative path of the note.",
            "title": "Id",
            "type": "string"
          },
          "modified_at": {
            "description": "File mtime in the configured timezone.",
            "format": "date-time",
            "title": "Modified At",
            "type": "string"
          },
          "path": {
            "description": "Vault-relative path of the note.",
            "title": "Path",
            "type": "string"
          },
          "title": {
            "description": "Frontmatter `title`, else the file name without .md.",
            "title": "Title",
            "type": "string"
          },
          "truncated": {
            "description": "True when the note exceeded MAX_NOTE_SIZE_BYTES and content was cut.",
            "title": "Truncated",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "path",
          "title",
          "frontmatter",
          "content",
          "modified_at",
          "truncated"
        ],
        "title": "NoteResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "description": "Send `Authorization: Bearer <API_TOKEN>`.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Read-mostly gateway over an Obsidian vault: full-vault search, note reads, vault tree/summary browsing, inbox duplicate-candidate detection, and note creation and append restricted to 00_Inbox/ChatGPT. MCP (mounted at /mcp) is the primary interface; this REST API is kept for health checks, curl-based diagnostics, and regression tests. Every failing response uses the single `{\"error\": {\"code\": ..., \"message\": ...}}` envelope described on each operation below — never FastAPI's default `{\"detail\": ...}` shape.",
    "title": "Obsidian Vault Gateway",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/inbox/duplicate-candidates": {
      "get": {
        "description": "Read-only: never creates, appends to, or otherwise changes any note.\n\n``keywords`` takes the same comma-separated shape ``/search``'s ``tags``\ndoes; splitting it into a list is this router's job (the MCP tool sends\na JSON array directly) — folding/deduping happens once, inside\n``app/services/duplicate_notes.py``, for both transports.",
        "operationId": "findDuplicateCandidates",
        "parameters": [
          {
            "in": "query",
            "name": "title",
            "required": true,
            "schema": {
              "maxLength": 300,
              "minLength": 1,
              "title": "Title",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "project",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Project"
            }
          },
          {
            "description": "Comma-separated keywords, matched against title/tags only.",
            "in": "query",
            "name": "keywords",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Comma-separated keywords, matched against title/tags only.",
              "title": "Keywords"
            }
          },
          {
            "description": "Maximum candidates to return.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 5,
              "description": "Maximum candidates to return.",
              "maximum": 10,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DuplicateCandidatesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`error.code` is one of: INVALID_FILE_TYPE, VALIDATION_ERROR."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`error.code` is one of: UNAUTHORIZED."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`error.code` is one of: NOTE_NOT_FOUND."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`error.code` is one of: INTERNAL_ERROR."
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Every failing response — including any status code not listed individually above — uses this same envelope; only `error.code` and `error.message` vary."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Find 00_Inbox/ChatGPT notes that may duplicate a proposed title",
        "tags": [
          "inbox"
        ]
      }
    },
    "/api/v1/inbox/notes": {
      "post": {
        "operationId": "createInboxNote",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InboxNoteCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedNoteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`error.code` is one of: INVALID_TITLE, VALIDATION_ERROR."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`error.code` is one of: UNAUTHORIZED."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`error.code` is one of: NOTE_ALREADY_EXISTS."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`error.code` is one of: FILE_TOO_LARGE."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`error.code` is one of: INTERNAL_ERROR."
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Every failing response — including any status code not listed individually above — uses this same envelope; only `error.code` and `error.message` vary."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Create a Markdown note in 00_Inbox/ChatGPT",
        "tags": [
          "inbox"
        ]
      }
    },
    "/api/v1/inbox/notes/append": {
      "post": {
        "operationId": "appendInboxNote",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InboxNoteAppendRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppendedNoteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`error.code` is one of: INVALID_PATH, INVALID_FILE_TYPE, VALIDATION_ERROR."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`error.code` is one of: UNAUTHORIZED."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`error.code` is one of: PATH_OUTSIDE_VAULT."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`error.code` is one of: NOTE_NOT_FOUND."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`error.code` is one of: NOTE_MODIFIED."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`error.code` is one of: FILE_TOO_LARGE."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`error.code` is one of: INTERNAL_ERROR."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`error.code` is one of: INBOX_LOCK_TIMEOUT."
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Every failing response — including any status code not listed individually above — uses this same envelope; only `error.code` and `error.message` vary."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Append Markdown to an existing note in 00_Inbox/ChatGPT",
        "tags": [
          "inbox"
        ]
      }
    },
    "/api/v1/notes": {
      "get": {
        "operationId": "readNote",
        "parameters": [
          {
            "description": "Vault-relative path, e.g. 'Knowledge/PC/GPU/RTX 5070.md'.",
            "in": "query",
            "name": "path",
            "required": true,
            "schema": {
              "description": "Vault-relative path, e.g. 'Knowledge/PC/GPU/RTX 5070.md'.",
              "title": "Path",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`error.code` is one of: INVALID_PATH, INVALID_FILE_TYPE, VALIDATION_ERROR."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`error.code` is one of: UNAUTHORIZED."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`error.code` is one of: PATH_OUTSIDE_VAULT."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`error.code` is one of: NOTE_NOT_FOUND."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`error.code` is one of: INTERNAL_ERROR."
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Every failing response — including any status code not listed individually above — uses this same envelope; only `error.code` and `error.message` vary."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Read a note by its vault-relative path",
        "tags": [
          "notes"
        ]
      }
    },
...
  }
}